small typo's and some small adjustments
Line 107  : We had 9 origin points  =  We had 13 origin points  (13 origins not 9 in the file !)
Line 142  : click on the origin point (OBJECTID_1 = 853046) in the canvas  == This point might not be in the resultset as the points are randomly generated !
Line 142  : Topology tolerance as 0.0000150  -- (image says 0.000015)   => number of decimals seems off!   -and-  
Line 157  :  In Topology tollerance enter 0.000015. --  removed one "l"  (image says 0.000150)    => number of decimals seems off!
Line 167  :   In the :guilabel:`History` dialog, Select the lastest command(command used for Shortest path).  == should probably be
                   In the :guilabel:`History` dialog, select the latest command (command used for Shortest path).  = lowercase s (select)   remove s (lastest)     space after "command"
Line 199  :  A new layer ``Network Routes`` Layer will be added to the  = A new layer ``Network Routes`` will be added to the   ===   removed "Layer"