Some minor changes
Line 20 : URL of the Public laboratory seems to have changed from HTTP to HTTPS and also the name changed to https://publiclab.org
Line 22 : URL of the Public laboratory seems to have changed from HTTP to HTTPS and also the name changed to https://publiclab.org. also removed one superfluous space
Line 40 : "But that is cumbersome and practical. "should probably be : "But that is cumbersome and not practical." Added "not" to the sentence
Line 65 : "default ad choose ``LZW`` should probably be "default and choose ``LZW``. Changed "ad" to and"